# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OllamaModel {
  name: string;
}

export interface OllamaModelCatalog {
  models: OllamaModel[];
}

export type BookSourceType = typeof BookSourceType[keyof typeof BookSourceType];


export const BookSourceType = {
  EPUB: 'EPUB',
  ARTICLE: 'ARTICLE',
} as const;

export type BookStatus = typeof BookStatus[keyof typeof BookStatus];


export const BookStatus = {
  READY: 'READY',
  PROCESSING: 'PROCESSING',
  DRAFT: 'DRAFT',
} as const;

export interface Book {
  id: string;
  title: string;
  author: string;
  sourceType: BookSourceType;
  status: BookStatus;
  level: string;
  progress: number;
  /** Last chapter opened */
  readingChapter: number;
  /** Last offset in words */
  readingOffset: number;
  coverColor: string;
  updatedAt: string;
}

export interface StudyItem {
  term: string;
  meaning: string;
  example: string;
  pronunciation: string;
  difficulty: string;
}

export interface VisualStudyCard {
  term: string;
  meaning: string;
  example: string;
  visualCue: string;
  technique: string;
  pronunciation: string;
  difficulty: string;
}

export interface LinguisticDeck {
  id: string;
  title: string;
  purpose: string;
  items: StudyItem[];
}

export interface SemanticNode {
  id: string;
  label: string;
  description: string;
}

export interface SemanticConnection {
  fromId: string;
  toId: string;
  relationship: string;
}

export interface SemanticMap {
  nodes: SemanticNode[];
  connections: SemanticConnection[];
}

export interface StudyPlan {
  vocabulary: StudyItem[];
  idioms: StudyItem[];
  phrasalVerbs: StudyItem[];
  visualCards?: VisualStudyCard[];
  linguisticDecks?: LinguisticDeck[];
  semanticMap?: SemanticMap;
}

export interface Chapter {
  id: string;
  /** Chapter order */
  position: number;
  title: string;
  content: string;
  /** Number of words in the chapter */
  wordCount: number;
}

export type BookDetail = Book & {
  plan: StudyPlan;
  chapters: Chapter[];
};

export interface ReadingPositionRequest {
  /** @minimum 1 */
  chapter: number;
  /** @minimum 0 */
  offset: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
}

export type ImportBookRequestSourceType = typeof ImportBookRequestSourceType[keyof typeof ImportBookRequestSourceType];


export const ImportBookRequestSourceType = {
  EPUB: 'EPUB',
  ARTICLE: 'ARTICLE',
} as const;

/**
 * Provider used to prepare the plan.
 */
export type ImportBookRequestProvider = typeof ImportBookRequestProvider[keyof typeof ImportBookRequestProvider];


export const ImportBookRequestProvider = {
  ollama: 'ollama',
  openrouter: 'openrouter',
} as const;

export interface ImportBookRequest {
  title: string;
  author: string;
  sourceType: ImportBookRequestSourceType;
  /** For EPUB, a base64 data URL containing the complete file. For ARTICLE, the article URL. */
  content: string;
  /** Original EPUB file name, used to validate the upload. */
  fileName?: string;
  /** Local Ollama base URL. */
  ollamaEndpoint: string;
  /** Model used to prepare the plan. */
  ollamaModel: string;
  /** Provider used to prepare the plan. */
  provider: ImportBookRequestProvider;
}

export interface DictionaryImportRequest {
  title: string;
  publisher?: string;
  isbn?: string;
  fileName: string;
  /** Base64 data URL with the private EPUB file. */
  content: string;
  privateAcknowledged: boolean;
}

export interface DictionarySourceSummary {
  id: string;
  title: string;
  publisher: string;
  isbn: string;
  /** Number of parsed entries */
  entryCount: number;
  createdAt: string;
}

export interface DictionaryImportResult {
  source: DictionarySourceSummary;
  importedEntries: number;
  skippedLines: number;
  warnings: string[];
}

export interface DictionaryEntrySummary {
  id: string;
  headword: string;
  term: string;
  translation: string;
  partOfSpeech: string;
  usageLabels: string[];
  sourceTitle: string;
  exampleCount: number;
}

export interface DictionarySense {
  id: string;
  definition: string;
  translation: string;
}

export interface DictionaryExample {
  id: string;
  sentence: string;
  translation: string;
  explanation: string;
  createdAt: string;
}

export interface DictionaryStudyCard {
  id: string;
  headword?: string;
  term: string;
  translation: string;
  /** @nullable */
  exampleId?: string | null;
}

export type DictionaryEntryDetailSource = {
  title: string;
  publisher: string;
};

export interface DictionaryEntryDetail {
  id: string;
  term: string;
  translation: string;
  partOfSpeech: string;
  usageLabels: string[];
  source: DictionaryEntryDetailSource;
  senses: DictionarySense[];
  examples: DictionaryExample[];
  cards: DictionaryStudyCard[];
}

export interface PublicDictionaryEntrySummary {
  id: string;
  term: string;
  partOfSpeech: string;
  senseCount: number;
}

export interface PublicDictionarySense {
  id: string;
  definition: string;
  position: number;
}

export interface PublicDictionaryForm {
  id: string;
  form: string;
  tags: string;
}

export interface PublicDictionarySound {
  id: string;
  ipa: string;
  audioUrl: string;
}

export interface PublicDictionarySource {
  name: string;
  version: string;
  license: string;
  attribution: string;
}

export interface PublicDictionaryEntryDetail {
  id: string;
  term: string;
  partOfSpeech: string;
  senses: PublicDictionarySense[];
  forms: PublicDictionaryForm[];
  sounds: PublicDictionarySound[];
  source: PublicDictionarySource;
}

export type DictionaryExampleRequestProvider = typeof DictionaryExampleRequestProvider[keyof typeof DictionaryExampleRequestProvider];


export const DictionaryExampleRequestProvider = {
  ollama: 'ollama',
} as const;

export interface DictionaryExampleRequest {
  provider: DictionaryExampleRequestProvider;
  endpoint: string;
  model: string;
}

export interface DictionaryCardRequest {
  exampleId?: string;
}

export interface Dashboard {
  minutesToday: number;
  streak: number;
  wordsLearned: number;
  currentBook: Book;
}

export type ListOllamaModelsParams = {
/**
 * Base URL of the local Ollama instance.
 */
endpoint: string;
};

export type ListOpenRouterModelsParams = {
/**
 * OpenRouter API base URL.
 */
endpoint?: string;
};

export type SearchDictionaryEntriesParams = {
query?: string;
};

export type SearchPublicDictionaryParams = {
query?: string;
/**
 * Maximum results
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};