# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public-dictionary:
    get:
      operationId: searchPublicDictionary
      tags: [study]
      summary: Search the imported open English dictionary
      parameters:
        - name: query
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: number, minimum: 1, maximum: 100, default: 40, description: Maximum results }
      responses:
        "200":
          description: Matching public dictionary entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PublicDictionaryEntrySummary" }
  /public-dictionary/{entryId}:
    get:
      operationId: getPublicDictionaryEntry
      tags: [study]
      summary: Get an entry from the imported open dictionary
      parameters:
        - $ref: "#/components/parameters/DictionaryEntryId"
      responses:
        "200":
          description: Public dictionary entry with senses, forms, pronunciation and attribution
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicDictionaryEntryDetail" }
...
components:
  parameters:
...
    DictionaryEntryId:
      name: entryId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    PublicDictionaryEntrySummary:
      type: object
      properties:
        id: { type: string }
        term: { type: string }
        partOfSpeech: { type: string }
        senseCount: { type: number, integer: true }
      required: [id, term, partOfSpeech, senseCount]
    PublicDictionarySense:
      type: object
      properties:
        id: { type: string }
        definition: { type: string }
        position: { type: number, integer: true }
      required: [id, definition, position]
    PublicDictionaryForm:
      type: object
      properties:
        id: { type: string }
        form: { type: string }
        tags: { type: string }
      required: [id, form, tags]
    PublicDictionarySound:
      type: object
      properties:
        id: { type: string }
        ipa: { type: string }
        audioUrl: { type: string }
      required: [id, ipa, audioUrl]
    PublicDictionarySource:
      type: object
      properties:
        name: { type: string }
        version: { type: string }
        license: { type: string }
        attribution: { type: string }
      required: [name, version, license, attribution]
    PublicDictionaryEntryDetail:
      type: object
      properties:
        id: { type: string }
        term: { type: string }
        partOfSpeech: { type: string }
        senses:
          type: array
          items: { $ref: "#/components/schemas/PublicDictionarySense" }
        forms:
          type: array
          items: { $ref: "#/components/schemas/PublicDictionaryForm" }
        sounds:
          type: array
          items: { $ref: "#/components/schemas/PublicDictionarySound" }
        source: { $ref: "#/components/schemas/PublicDictionarySource" }
      required: [id, term, partOfSpeech, senses, forms, sounds, source]
...