# lib/api-zod/src/generated/types/publicDictionaryEntryDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PublicDictionaryForm } from './publicDictionaryForm';
import type { PublicDictionarySense } from './publicDictionarySense';
import type { PublicDictionarySound } from './publicDictionarySound';
import type { PublicDictionarySource } from './publicDictionarySource';

export interface PublicDictionaryEntryDetail {
  id: string;
  term: string;
  partOfSpeech: string;
  senses: PublicDictionarySense[];
  forms: PublicDictionaryForm[];
  sounds: PublicDictionarySound[];
  source: PublicDictionarySource;
}